Refactor code: move display function to actions class
